backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distlib==0.3.8
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
idna==3.10
oauthlib==3.2.2
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
